@@ -29,7 +29,7 @@
     "build:html": "cp public/*.html dist/.",
     "build:ico": "cp public/*.ico dist/.",
     "build:img": "cp public/*.png dist/public/.",
-    "start": "NODE_ENV=development webpack-dev-server --config webpack.config.js --progress --colors",
+    "start": "NODE_ENV=development webpack-dev-server --progress --colors",
     "start:heroku": "UNLEASH_API=http://unleash.herokuapp.com npm run start",
     "lint": "eslint . --ext js,jsx",
     "test": "jest",
@@ -59,30 +59,29 @@
   "devDependencies": {
     "babel-core": "^6.14.0",
     "babel-eslint": "^7.1.0",
-    "babel-loader": "^6.2.5",
+    "babel-loader": "^7.1.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-preset-es2015": "^6.14.0",
     "babel-preset-react": "^6.11.1",
     "babel-preset-react-hmre": "^1.1.1",
     "babel-preset-stage-0": "^6.5.0",
     "babel-preset-stage-2": "^6.13.0",
-    "css-loader": "^0.25.0",
+    "css-loader": "^0.28.4",
     "eslint": "^4.1.1",
     "eslint-config-finn": "^2.0.0",
     "eslint-config-finn-react": "^2.0.0",
     "eslint-plugin-react": "^7.1.0",
-    "extract-text-webpack-plugin": "^1.0.1",
+    "extract-text-webpack-plugin": "^3.0.0-rc.0",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^20.0.4",
-    "node-sass": "~3.12.1",
-    "postcss-loader": "^0.13.0",
+    "node-sass": "^4.5.3",
     "react-test-renderer": "^15.4.2",
     "redux-devtools": "^3.3.1",
-    "sass-loader": "^4.0.2",
-    "style-loader": "^0.13.1",
+    "sass-loader": "^6.0.6",
+    "style-loader": "^0.18.2",
     "toolbox-loader": "0.0.3",
-    "webpack": "^1.13.2",
-    "webpack-dev-server": "^1.15.1"
+    "webpack": "^3.1.0",
+    "webpack-dev-server": "^2.5.1"
   },
   "jest": {
     "moduleNameMapper": {
